+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 * Day report with horizontal timelines over projects as rows: `track report day [DATE]` (#25)
+* Command `track break DURATION` to insert breaks (#28)
 
 ### Bug fixes